Next Generation Clustered Heat Map (NGCHM) Construction Library

NGCHM provides tools for defining the contents of a new NGCHM, and for compiling and installing it on a NGCHM server.

Details

Typical usage (see example) is to create a base NGCHM using chmNew; extend it with at least one ngchmLayer; typically extend it further with an additional ngchmLayer, row and column dendrograms, classification bars, and popup menu entries; compile and install it on an available ngchmServer.

Initialization

When first loaded the NGCHM library reads configuration files in the directories specified by the NGCHMCONFIGPATH environment variable. This is a colon (:) separated list of directory names. If not set it defaults to /etc/ngchm:/usr/local/ngchm:/opt/ngchm:$HOME/.ngchm. See NGCHM-initialization for details.

See also

Examples

data(TCGA.GBM.EXPR)
chm1 <- chmNew('gbm', TCGA.GBM.EXPR[1:50,1:50], rowAxisType='bio.gene.hugo', colAxisType='bio.tcga.barcode.sample.vial.portion.analyte.aliquot');
if (FALSE) chmExportToFile(chm1, 'gbm.ngchm');
chmExportToPDF(chm1, 'gbm.pdf');

#> Error in chmExportToPDF(chm1, "gbm.pdf"): shaidyMapGen required but not specified or set in environment

mat <- matrix(rnorm(100),nrow=10)
rownames(mat) <- sprintf ("ABCA%d", 1:10)
colnames(mat) <- sprintf ("Sample%d", 1:10)
chm <- chmNew ('my-chm', mat)
if (FALSE) chmSetCollection ('//server/collection')
chmInstall (chm)

#> Error in .local(chm, ...): length(dest$server) > 0 is not TRUE