Get information about a type name.

This function gets any registered information about a type name used for determining row and column linkouts. Registration of a typename is (currently) not required in order to use it, so it's possible for valid type name not to have any registered information.

chmGetTypeInfo(typename)

Arguments

typename
The name of the type.

Value

A list containing basic information about the type.

See also