Create a new NGCHM.

This function creates a Next Generation Clustered Heat Map (NGCHM) object in memory. Additional parameters will be added to the new NGCHM (see chmAdd). The bare NGCHM needs at least one data layer added to it before it can be compiled. This function requires git to be installed.

Usage

chmNew(
  name,
  ...,
  rowOrder = chmDefaultRowOrder,
  rowDist = "correlation",
  rowAgglom = "ward.D2",
  colOrder = chmDefaultColOrder,
  colDist = "correlation",
  colAgglom = "ward.D2",
  rowAxisType = NULL,
  colAxisType = NULL,
  rowCovariates = NULL,
  colCovariates = NULL,
  format = "original",
  rowGapLocations = NULL,
  rowGapWidth = 5,
  colGapLocations = NULL,
  colGapWidth = 5,
  overview = c(),
  logLevel = "INFO",
  logFile = NULL
)

Arguments

name

The name under which the NGCHM will be saved to the NGCHM server.

...

Zero or more initial objects to include in the NGCHM (see chmAdd).

rowOrder

A vector, dendrogram, or function specifying the CHM row order.

rowDist

Distance method to use by default RowOrder

rowAgglom

Agglomeration method to use by default RowOrder

colOrder

A vector, dendrogram, or function specifying the CHM column order.

colDist

Distance method to use by default ColOrder

colAgglom

Agglomeration method to use by default ColOrder

rowAxisType

The type(s) of the row labels (default: None).

colAxisType

The type(s) of the column labels (default: None).

rowCovariates

Covariate(Bar)(s) to add to the rows (default: None).

colCovariates

Covariate(Bar)(s) to add to the columns (default: None).

format

The format of NGCHM to produce (default: 'original').

rowGapLocations

Locations for row gaps. Specify as a list of integers or chmTreeGaps() function.

rowGapWidth

Width of row gaps (default: 5 rows)

colGapLocations

Locations for col gaps. Specify as a list of integers or chmTreeGaps() function.

colGapWidth

Width of col gaps (default: 5 cols)

overview

The format(s) of overview image(s) to create (default: None).

logLevel

The level of logs to output

logFile

The file to which logs should be output

Value

An object of class ngchm

See also

Examples

mychm <- chmNew("test_chm")
mychm <- chmNew("test_chm", rowGapLocations = c(3, 5))
mychm <- chmNew("test_chm", rowGapLocations = chmTreeGaps(4))
mychm <- chmNew("test_chm", rowGapWidth = 3)